Reorganization of Repo
Reorganized repo for main and desktop
